完成活动模块，准备重构用户模块

+++ app/Api/Controllers/ActivityController.php
@@ -10,12 +10,19 @@ namespace App\Api\Controllers;
 
 
 use App\Model\Activity;
+use App\Model\Tag;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Validator;
 use JWTAuth;
 
 class ActivityController extends BaseController
 {
+    public function create()
+    {
+        $tags=Tag::lists('name','id');
+        return compact('tags');
+    }
+
     public function index()
     {
         $activities=Activity::with('creator')->get();
@@ -36,6 +43,10 @@ class ActivityController extends BaseController
     public function show($id)
     {
         $activity=Activity::with('creator')->find($id);
+        if($activity==null)
+        {
+            return $this->return_response_activity('4004','未找到相关信息');
+        }
         $activity=$this->insert_tags($activity);
         return $this->return_response_activity('2000','success',$activity);
     }
@@ -61,12 +72,20 @@ class ActivityController extends BaseController
         }
         $activity=Activity::with('creator')->find($activity->id);
         $activity=$this->insert_tags($activity);
-        return $this->return_response_activity('2000','success',$activity);
+        return $this->return_response_activity('2000','发布成功',$activity);
     }
 
     public function update($id)
     {
         $activity=Activity::with('creator')->find($id);
+        if($activity==null)
+        {
+            return $this->return_response_activity('4004','未找到相关信息');
+        }
+        if($activity->creator!=$this->getUser()->id)
+        {
+            return $this->return_response_activity('4030','您没有权限进行此操作');
+        }
         $rules = [
             'title'   => 'required|max:100',
             'content' => 'required',
@@ -84,7 +103,7 @@ class ActivityController extends BaseController
             return $this->return_response_activity('5000','服务器出错');
         }
         $activity=$this->insert_tags($activity);
-        return $this->return_response_activity('2000','success',$activity);
+        return $this->return_response_activity('2000','更新成功',$activity);
     }
 
     public function destroy($id)
@@ -92,15 +111,19 @@ class ActivityController extends BaseController
         $activity=Activity::find($id);
         if($activity==null)
         {
-            return $this->return_response_activity('5000','未找到相关信息',(int)$id);
+            return $this->return_response_activity('4004','未找到相关信息');
+        }
+        if($activity->creator!=$this->getUser()->id)
+        {
+            return $this->return_response_activity('4030','您没有权限进行此操作');
         }
         if(!$activity->delete())
         {
-            return $this->return_response_activity('5000','服务器出错',(int)$id);
+            return $this->return_response_activity('5000','服务器出错');
         }
         else
         {
-            return $this->return_response_activity('2000','删除成功',(int)$id);
+            return $this->return_response_activity('2000','删除成功');
         }
     }
 }
